refactor(transferprog): share put.io error-string mapping across RPC + dashboard
The dashboard's render.go carried a literal copy of torrent.go's error
precedence ("copy torrent.go:handleTorrentGet") that only mapped
ErrorMessage, so the #22 tracker/status-message handling would have applied
to *arr but not the web dashboard. Move PutioErrorString into transferprog
— the package that already exists to keep the RPC server (:9091) and the
dashboard reporting identical numbers — and call it from both faces.

No behavior change for the RPC path; the dashboard now surfaces the same
tracker > error > status reason that *arr sees.
